+++ truffle/src/com.oracle.truffle.api.dsl.test/src/com/oracle/truffle/api/dsl/test/ImplicitCastTest.java
@@ -42,10 +42,10 @@ package com.oracle.truffle.api.dsl.test;
 
 import static com.oracle.truffle.api.dsl.test.examples.ExampleNode.createArguments;
 
-import java.lang.reflect.Field;
 import java.util.concurrent.locks.ReentrantLock;
 
 import org.junit.Assert;
+import org.junit.Assume;
 import org.junit.Test;
 
 import com.oracle.truffle.api.CallTarget;
@@ -649,17 +649,17 @@ public class ImplicitCastTest {
     }
 
     @Test
-    public void test33Bits() throws NoSuchFieldException, SecurityException {
+    public void test33Bits() {
+        Assume.assumeTrue(StateBitTest.getStateBitWidth() >= 32);
         ExampleNode node = ThirtyThreeBitsNodeGen.create(null);
-        Field stateField = node.getClass().getDeclaredField("state_");
-        Assert.assertEquals(long.class, stateField.getType());
+        StateBitTest.assertStateFields(node, 33);
     }
 
     @Test
-    public void test32Bits() throws NoSuchFieldException, SecurityException {
+    public void test32Bits() {
+        Assume.assumeTrue(StateBitTest.getStateBitWidth() >= 32);
         ExampleNode node = ThirtyTwoBitsNodeGen.create(null);
-        Field stateField = node.getClass().getDeclaredField("state_");
-        Assert.assertEquals(int.class, stateField.getType());
+        StateBitTest.assertStateFields(node, 32);
     }
 
     @TypeSystem