+++ truffle/CHANGELOG.md
@@ -13,6 +13,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 * Added `HostCompilerDirectives` for directives that guide the host compilations of Truffle interpreters.
     * `HostCompilerDirectives.BytecodeInterpreterSwitch` - to denote methods that contain the instruction-dispatch switch in bytecode interpreters
     * `HostCompilerDirectives.BytecodeInterpreterSwitchBoundary` - to denote methods that do not need to be inlined into the bytecode interpreter switch
+* Truffle DSL generated nodes are no longer limited to 64 state bits. Use these state bits responsibly.
 
 ## Version 20.3.0
 * Added `RepeatingNode.initialLoopStatus` and `RepeatingNode.shouldContinue` to allow defining a custom loop continuation condition.